Solving the mystery of the missing index.html in commontechdoc
git-svn-id: https://gtsvn.uit.no/langtech/trunk/giella-core@178812 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

+++ scripts/static-divvun.py
@@ -328,7 +328,7 @@ class StaticSiteBuilder(object):
 
     def copy_to_site(self):
         """Copy the entire site to self.destination."""
-        if 'techdoc' in self.builddir:
+        if 'techdoc' in self.builddir and 'commontec' not in self.builddir:
             offending_file = os.path.join(self.builddir, 'built', 'index.html')
             if os.path.exists(offending_file):
                 os.remove(offending_file)
